Add card filtering to NOfAKind and TwoPair classes
This allows us to pass in the full hand to their initializers and they
will automatically extract the necessary cards. This allows our Parser
to easily generate the remainder HighCard group if the valid group has
fewer than 5 cards.

Also added the .within? method to easily check if a given card group is
contained within a set of cards.