Fix high card / remainder scoring
Resolves #1

Previously, the score for a High Card / remainder of a hand was simply
the high value of the card. This neglected potential ties of the highest
card (which weren't tested in hand_evaluator_spec.rb. We now calculate a
decimal corresponding to the ordered values of the cards in the group,
which properly handles ties. Credit to @jwrobes for inspiration.